landmark delete check

+++ dotnet/portalserver/net.atos.daf.ct2.portalservice/Controllers/LandmarkCorridorController.cs
@@ -10,6 +10,10 @@ using net.atos.daf.ct2.corridorservice;
 using net.atos.daf.ct2.portalservice.Common;
 using net.atos.daf.ct2.portalservice.Entity.Corridor;
 using Newtonsoft.Json;
+using System;
+using System.Reflection;
+using System.Threading.Tasks;
+using Alert = net.atos.daf.ct2.alertservice;    
 
 namespace net.atos.daf.ct2.portalservice.Controllers
 {
@@ -23,13 +27,14 @@ namespace net.atos.daf.ct2.portalservice.Controllers
         private readonly CorridorService.CorridorServiceClient _corridorServiceClient;
         private readonly AuditHelper _auditHelper;
         private readonly CorridorMapper _corridorMapper;
-
-        public LandmarkCorridorController(CorridorService.CorridorServiceClient corridorServiceClient, AuditHelper auditHelper, AccountPrivilegeChecker privilegeChecker, IHttpContextAccessor _httpContextAccessor, SessionHelper sessionHelper) : base(_httpContextAccessor, sessionHelper, privilegeChecker)
+        private readonly Alert.AlertService.AlertServiceClient _alertServiceClient;
+        public LandmarkCorridorController(CorridorService.CorridorServiceClient corridorServiceClient, AuditHelper auditHelper, Common.AccountPrivilegeChecker privilegeChecker, Alert.AlertService.AlertServiceClient alertServiceClient, IHttpContextAccessor _httpContextAccessor, SessionHelper sessionHelper) : base(_httpContextAccessor, sessionHelper)
         {
             _logger = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
             _corridorServiceClient = corridorServiceClient;
             _auditHelper = auditHelper;
             _corridorMapper = new CorridorMapper();
+            _alertServiceClient = alertServiceClient;
         }
 
         [HttpGet]
@@ -251,8 +256,20 @@ namespace net.atos.daf.ct2.portalservice.Controllers
                 {
                     return StatusCode(400, "Corridor id is required.");
                 }
+
+                Alert.LandmarkIdRequest landmarkIdRequest = new Alert.LandmarkIdRequest();
+                landmarkIdRequest.LandmarkId.Add(request.Id);
+
+                Alert.LandmarkIdExistResponse isLandmarkavalible = await _alertServiceClient.IsLandmarkActiveInAlertAsync(landmarkIdRequest);
+
+                if (isLandmarkavalible.IsLandmarkActive)
+                {
+                    return StatusCode(409, "Corridor is used in alert.");
+                }
                 var MapRequest = _corridorMapper.MapId(request);
                 var data = await _corridorServiceClient.DeleteCorridorAsync(MapRequest);
+
+
                 if (data != null && data.Code == Responsecode.Success)
                 {
                     await _auditHelper.AddLogs(DateTime.Now, "Landmark Corridor Component",
